[NU-1979] FlinkDeploymentManager simpler - flatten 3 levels of inheritance
