+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/periodic/flink/FlinkClientStub.scala
@@ -8,24 +8,31 @@ import pl.touk.nussknacker.engine.management.rest.{FlinkClient, flinkRestModel}
 import java.io.File
 import scala.concurrent.Future
 
-class FlinkClientStub extends FlinkClient {
+object FlinkClientStub extends FlinkClient {
+
+  val maxParallelism    = 10
+  val savepointPath     = "savepoints/123-savepoint"
+  val stopSavepointPath = "savepoints/246-stop-savepoint"
 
   override def getJobsOverviews()(
       implicit freshnessPolicy: DataFreshnessPolicy
-  ): Future[WithDataFreshnessStatus[List[flinkRestModel.JobOverview]]] = ???
+  ): Future[WithDataFreshnessStatus[List[flinkRestModel.JobOverview]]] =
+    Future.successful(WithDataFreshnessStatus.fresh(List.empty))
 
-  override def getJobDetails(jobId: String): Future[Option[flinkRestModel.JobDetails]] = ???
+  override def getJobDetails(jobId: String): Future[Option[flinkRestModel.JobDetails]] = Future.successful(None)
 
-  override def getJobConfig(jobId: String): Future[flinkRestModel.ExecutionConfig] = ???
+  override def getJobConfig(jobId: String): Future[flinkRestModel.ExecutionConfig] =
+    Future.successful(flinkRestModel.ExecutionConfig(1, Map.empty))
 
-  override def cancel(deploymentId: ExternalDeploymentId): Future[Unit] = ???
+  override def cancel(deploymentId: ExternalDeploymentId): Future[Unit] = Future.successful(())
 
   override def makeSavepoint(
       deploymentId: ExternalDeploymentId,
       savepointDir: Option[String]
-  ): Future[SavepointResult] = ???
+  ): Future[SavepointResult] = Future.successful(SavepointResult(savepointPath))
 
-  override def stop(deploymentId: ExternalDeploymentId, savepointDir: Option[String]): Future[SavepointResult] = ???
+  override def stop(deploymentId: ExternalDeploymentId, savepointDir: Option[String]): Future[SavepointResult] =
+    Future.successful(SavepointResult(stopSavepointPath))
 
   override def runProgram(
       jarFile: File,
@@ -33,12 +40,15 @@ class FlinkClientStub extends FlinkClient {
       args: List[String],
       savepointPath: Option[String],
       jobId: Option[String]
-  ): Future[Option[ExternalDeploymentId]] = ???
+  ): Future[Option[ExternalDeploymentId]] = Future.successful(None)
 
   override def deleteJarIfExists(jarFileName: String): Future[Unit] = Future.successful(())
 
-  override def getClusterOverview: Future[flinkRestModel.ClusterOverview] = ???
+  override def getClusterOverview: Future[flinkRestModel.ClusterOverview] =
+    Future.successful(
+      flinkRestModel.ClusterOverview(`slots-total` = maxParallelism, `slots-available` = maxParallelism)
+    )
 
-  override def getJobManagerConfig: Future[Configuration] = ???
+  override def getJobManagerConfig: Future[Configuration] = Future.successful(new Configuration)
 
 }
