+++ engine/flink/management/src/main/resources/META-INF/services/pl.touk.nussknacker.engine.DeploymentManagerProvider
@@ -1,1 +1,1 @@
-pl.touk.nussknacker.engine.management.FlinkStreamingDeploymentManagerProvider
+pl.touk.nussknacker.engine.management.FlinkDeploymentManagerProvider
